+++ R/bi_scale_fill.R
@@ -4,13 +4,15 @@
 #'     is used and the \code{bi_class} variable is given as the \code{fill} in the aesthetic
 #'     mapping.
 #'
-#' @usage bi_scale_fill(pal, dim = 3, ...)
+#' @usage bi_scale_fill(pal, dim = 3, flip_axes = FALSE, rotate_pal = FALSE, ...)
 #'
 #' @param pal Either palette name (one of \code{"Brown"}, \code{"DkBlue"},
 #'     \code{"DkCyan"}, \code{"DkViolet"}, or \code{"GrPink"}) or a custom palette
 #'     object created with \code{\link{bi_pal_manual}}.
 #' @param dim The dimensions of the palette, either \code{2} for a two-by-two palette or
 #'     \code{3} for a three-by-three palette.
+#' @param flip_axes A logical scalar; if \code{TRUE} (default: FALSE) the axes of the palette will be flipped.
+#' @param rotate_pal A logical scalar; if \code{TRUE} (default: FALSE) the palette will be rotated 180 degrees.
 #' @param ... Arguments to pass to \code{\link{scale_fill_manual}}
 #'
 #' @return A \code{ggplot} object with the given bivariate palette applied to the data.
@@ -39,12 +41,11 @@
 #'   bi_scale_fill(pal = "GrPink", dim = 3)
 #'
 #' @export
-bi_scale_fill <- function(pal, dim = 3, ...){
+bi_scale_fill <- function(pal, dim = 3, flip_axes = FALSE, rotate_pal = FALSE, ...){
 
   # check parameters
   if (missing(pal) == TRUE){
-    stop("A palette must be specified for the 'pal' argument. Please choose one of: 'Brown', 'DkBlue', 'DkCyan', 'DkViolet', or 'GrPink'
-         or supply a custom palette created with 'bi_pal_custom()'.")
+    stop("A palette must be specified for the 'pal' argument. Please choose one of: 'BlGold', 'BlOrange', 'BlYellow', 'Brown', 'Diverging', 'DkBlue', 'DkCyan', 'DkViolet', 'Fire', 'GnPink', 'GnPurple', 'GrPink', 'OrgPurple', 'Reds' or 'Viridis' or supply a custom palette created with 'bi_pal_custom()'.")
   }
 
   if ("bi_pal_custom" %in% class(pal) == TRUE) {
@@ -57,8 +58,8 @@ bi_scale_fill <- function(pal, dim = 3, ...){
 
   } else if ("bi_pal_custom" %in% class(pal) == FALSE){
 
-    if (pal %in% c("Brown", "DkBlue", "DkCyan", "DkViolet", "GrPink") == FALSE){
-      stop("The given palette is not one of the allowed options for bivariate mapping. Please choose one of: 'Brown', 'DkBlue', 'DkCyan', 'DkViolet', or 'GrPink'.")
+    if (pal %in% c("BlGold", "BlOrange", "BlYellow", "Brown", "Diverging", "DkBlue", "DkCyan", "DkViolet", "Fire", "GnPink", "GnPurple", "GrPink", "OrgPurple", "Reds", "Viridis") == FALSE){
+      stop("The given palette is not one of the allowed options for bivariate mapping. Please choose one of: 'BlGold', 'BlOrange', 'BlYellow', 'Brown', 'Diverging', 'DkBlue', 'DkCyan', 'DkViolet', 'Fire', 'GnPink', 'GnPurple', 'GrPink', 'OrgPurple', 'Reds' or 'Viridis'.")
     }
 
   }
@@ -78,16 +79,30 @@ bi_scale_fill <- function(pal, dim = 3, ...){
 
   } else if ("bi_pal_custom" %in% class(pal) == FALSE){
 
-    if (pal == "DkViolet"){
-      x <- pal_dkviolet(n = dim)
-    } else if (pal == "GrPink"){
-      x <- pal_grpink(n = dim)
-    } else if (pal == "DkBlue"){
-      x <- pal_dkblue(n = dim)
-    } else if (pal == "DkCyan"){
-      x <- pal_dkcyan(n = dim)
-    } else if (pal == "Brown"){
-      x <- pal_brown(n = dim)
+    x <- switch(pal,
+      "DkViolet" = pal_dkviolet(n = dim),
+      "GrPink" = pal_grpink(n = dim),
+      "DkBlue" = pal_dkblue(n = dim),
+      "DkCyan" = pal_dkcyan(n = dim),
+      "Brown" = pal_brown(n = dim),
+      "BlGold" = pal_blgold(n = dim),
+      "BlOrange" = pal_blorange(n = dim),
+      "BlYellow" = pal_blyellow(n = dim),
+      "Viridis" = pal_viridis(n = dim),
+      "Diverging" = pal_diverging(n = dim),
+      "GnPink" = pal_gnpink(n = dim),
+      "GnPurple" = pal_gnpurp(n = dim),
+      "OrgPurple" = pal_orgpurp(n = dim),
+      "Fire" = pal_fire(n = dim),
+      "Reds" = pal_reds(n = dim)
+    )
+
+    if(flip_axes){
+      x <- bi_pal_flip(x)
+    }
+
+    if(rotate_pal){
+      x <- bi_pal_rotate(x)
     }
 
   }