+++ tests/testthat/test_bi_pal.R
@@ -13,12 +13,12 @@ library(sf)
 
 test_that("missing parameters trigger appropriate errors", {
   expect_error(bi_pal(dim = 3, preview = FALSE),
-               "A palette must be specified for the 'pal' argument. Please choose one of: 'Brown', 'DkBlue', 'DkCyan', 'DkViolet', and 'GrPink'.")
+               "A palette must be specified for the 'pal' argument. Please choose one of: 'BlGold', 'BlOrange', 'BlYellow', 'Brown', 'Diverging', 'DkBlue', 'DkCyan', 'DkViolet', 'Fire', 'GnPink', 'GnPurple', 'GrPink', 'OrgPurple', 'Reds' or 'Viridis'.")
 })
 
 test_that("incorrectly specified parameters trigger appropriate errors", {
   expect_error(bi_pal(pal = "ham", dim = 3, preview = FALSE),
-               "The given palette is not one of the allowed options for bivariate mapping. Please choose one of: 'Brown', 'DkBlue', 'DkCyan', 'DkViolet', and 'GrPink'.")
+               "The given palette is not one of the allowed options for bivariate mapping. Please choose one of: 'BlGold', 'BlOrange', 'BlYellow', 'Brown', 'Diverging', 'DkBlue', 'DkCyan', 'DkViolet', 'Fire', 'GnPink', 'GnPurple', 'GrPink', 'OrgPurple', 'Reds' or 'Viridis'.")
   expect_error(bi_pal(pal = "DkBlue", dim = 5, preview = FALSE),
                "The 'dim' argument only accepts the numeric values '2' or '3'.")
   expect_error(bi_pal(pal = "DkBlue", dim = "ham", preview = FALSE),