+++ tests/testthat/test_bi_scale_color.R
@@ -17,14 +17,14 @@ test_that("missing parameters trigger appropriate errors", {
   expect_error(ggplot() +
                  geom_sf(data = pointData, aes(color = bi_class), show.legend = FALSE) +
                  bi_scale_color(dim = 3),
-               "A palette must be specified for the 'pal' argument. Please choose one of: 'Brown', 'DkBlue', 'DkCyan', 'DkViolet', or 'GrPink'.")
+               "A palette must be specified for the 'pal' argument. Please choose one of: 'BlGold', 'BlOrange', 'BlYellow', 'Brown', 'Diverging', 'DkBlue', 'DkCyan', 'DkViolet', 'Fire', 'GnPink', 'GnPurple', 'GrPink', 'OrgPurple', 'Reds' or 'Viridis' or supply a custom palette created with 'bi_pal_custom()'.", fixed = TRUE)
 })
 
 test_that("incorrectly specified parameters trigger appropriate errors", {
   expect_error(ggplot() +
                  geom_sf(data = pointData, aes(color = bi_class), show.legend = FALSE) +
                  bi_scale_color(pal = "ham", dim = 2),
-               "The given palette is not one of the allowed options for bivariate mapping. Please choose one of: 'Brown', 'DkBlue', 'DkCyan', 'DkViolet', or 'GrPink'.")
+               "The given palette is not one of the allowed options for bivariate mapping. Please choose one of: 'BlGold', 'BlOrange', 'BlYellow', 'Brown', 'Diverging', 'DkBlue', 'DkCyan', 'DkViolet', 'Fire', 'GnPink', 'GnPurple', 'GrPink', 'OrgPurple', 'Reds' or 'Viridis'.")
   expect_error(ggplot() +
                  geom_sf(data = pointData, aes(color = bi_class), show.legend = FALSE) +
                  bi_scale_color(pal = "Brown", dim = 5),